+++ yii2-adapter/src/Event/EventCompatibility.php
@@ -28,6 +28,7 @@ use craft\services\Entries;
 use craft\services\Fields;
 use craft\services\Fs;
 use craft\services\Gc;
+use craft\services\Gql as LegacyGql;
 use craft\services\ImageTransforms;
 use craft\services\Plugins as LegacyPlugins;
 use craft\services\ProjectConfig as LegacyProjectConfig;
@@ -99,6 +100,7 @@ readonly class EventCompatibility
         Fields::registerEvents();
         Fs::registerEvents();
         Gc::registerEvents();
+        LegacyGql::registerEvents();
         LegacySearch::registerEvents();
         Utilities::registerEvents();
         Dashboard::registerEvents();
